requirements.txt
ShopifyAPI
orjson
jinja2
aiofiles
motor
fastapi-jwt-auth